Create OPENMPFibonacci.c
Write an OpenMP program to find and sum the Fibonacci series. Use one thread to generate the  numbers  up  to  the  specified  limit  N=50000  and  other  threads  has  to  sum  and  print them. Use omp reduction to protect the code region that might be updated by multiple threads concurrently. Measure the execution time for both versions (omp critical) varying the number of threads: 4 and 8.