Create MPISum.c
Write a MPI program to calculate the sum of numbers from 1 to 1000 in a parallel fashion while executing on all the cluster nodes and providing the result at the end on only one node. It should be noted that the print statement for the sum is only executed on the node that is ranked zero (O). Also calculate the execution time & wtime and find out environment variables. (Note: You have to use time stamp function to also print the time at begging and end of parallel code segment)