Create MPIArraySum.c
Write a MPI Program to perform: sending of arrays and simultaneous send and receive by array on each process is created, initialize it on process 0. Once the array has been initialized on process 0, then it is sent out to each process. Also calculate the execution time and time and find out environment variables.