Create MPIConwayGame.c
Conway’s Game of Life is played on a rectangular grid of cells that may or may not contain an organism. The state of the cells is updated each time step by applying the following set of rules:  
•  Every organism with two or three neighbours survives.
•  Every organism with four or more neighbours dies from overpopulation.
•  Every organism with zero or one neighbours dies from isolation.
•  Every empty cell adjacent to three organisms gives birth to a new one.
Create an MPI program that evolves a board of arbitrary size (dimensions could be specified at the command line) over several iterations. The board could be randomly generated or real from a file. Try applying the geometric decomposition pattern to partition the work among your process.