Create OPENMPParallelQuickSort.c
Using OpenMP, implement a Parallel Quick sort algorithm to sort a given set of elements and determine the time required to sort the elements. 
Repeat the experiment for different value of n, the number of elements in the list to be sorted. The elements can be read from a file or can be generated using the random number generator.